+++ tests/040_templatetags/045_categories.py
@@ -1,95 +1,114 @@
 import pytest
 
-from django.template import Context, Template
-from django.utils import translation
+from django.template import Context, Template, TemplateSyntaxError
 
 from lotus.factories import CategoryFactory
-from lotus.templatetags.lotus import article_categories
+from lotus.templatetags.lotus import get_categories
+from lotus.utils.tests import html_pyquery
 
 
-def assert_rendered_context(en_categories, rendered):
-    for category in en_categories:
-        assert {
-                "title": category.title,
-                "url": category.get_absolute_url()
-            } in rendered["categories"]
+def test_languages(db, rf):
+    """
+    Tag should only list categories for the current request language and possibly
+    flag item as "active" if a current category is given and match.
+    """
+    request = rf.get("/")
+
+    picsou = CategoryFactory(title="Picsou", slug="picsou", language="en")
+    CategoryFactory(title="Donald", slug="donald", language="en")
+    CategoryFactory(title="Dupont", slug="dupont", language="fr")
+
+    request.LANGUAGE_CODE = "en"
+    rendered = get_categories({"request": request}, current=picsou)
+    assert rendered == {
+        "categories": [
+            {
+                "title": "Donald",
+                "url": "/en/categories/donald/",
+                "is_active": False,
+            },
+            {
+                "title": "Picsou",
+                "url": "/en/categories/picsou/",
+                "is_active": True,
+            }
+        ]
+    }
+
+    request.LANGUAGE_CODE = "fr"
+    rendered = get_categories({"request": request}, current=picsou)
+    assert rendered == {
+        "categories": [
+            {
+                "title": "Dupont",
+                "url": "/fr/categories/dupont/",
+                "is_active": False,
+            },
+        ]
+    }
+
+    request.LANGUAGE_CODE = "de"
+    rendered = get_categories({"request": request})
+    assert rendered == {
+        "categories": []
+    }
 
 
-def test_english_language(db):
+def test_unavailable_language(db, rf):
+    """
+    Tag should not fail when current language is unknow, it just return empty result.
+    """
+    request = rf.get("/")
+    request.LANGUAGE_CODE = "zh"
 
-    with translation.override("en"):
-        en_categories = CategoryFactory.create_batch(3, language="en")
-        rendered = article_categories()
+    CategoryFactory(title="Picsou", slug="picsou", language="en")
 
-        assert len(rendered["categories"]) == 3
-        assert_rendered_context(en_categories, rendered)
-
-
-def test_french_language(db):
-    with translation.override("fr"):
-        fr_categories = CategoryFactory.create_batch(3, language="fr")
-        rendered = article_categories()
-
-        assert len(rendered["categories"]) == 3
-        assert_rendered_context(fr_categories, rendered)
+    rendered = get_categories({"request": request})
+    assert len(rendered["categories"]) == 0
 
 
-def test_no_categories(db):
-    with translation.override("en"):
-        rendered = article_categories()
+def test_current_invalid(db, rf):
+    """
+    Tag should raise a TemplateSyntaxError if given current argument is not a
+    Category object.
+    """
+    request = rf.get("/")
 
-        assert len(rendered["categories"]) == 0
+    CategoryFactory(title="Picsou", slug="picsou", language="en")
 
+    request.LANGUAGE_CODE = "en"
 
-def test_multiple_languages(db):
-    en_categories = CategoryFactory.create_batch(3, language="en")
-    fr_categories = CategoryFactory.create_batch(3, language="fr")
+    with pytest.raises(TemplateSyntaxError) as excinfo:
+        get_categories({"request": request}, current=42)
 
-    with translation.override("en"):
-        rendered = article_categories()
-        assert len(rendered["categories"]) == 3
-        assert_rendered_context(en_categories, rendered)
-
-    with translation.override("fr"):
-        rendered = article_categories()
-        assert len(rendered["categories"]) == 3
-        assert_rendered_context(fr_categories, rendered)
+    assert str(excinfo.value) == (
+        "'get_categories' tag only accepts a Category object as 'current' argument. "
+        "Object type 'int' was given."
+    )
 
 
-def test_unavailable_language(db):
-    CategoryFactory.create_batch(3, language="en")
+def test_template_rendering(db, rf):
+    """
+    HTML tag rendering test in-situ in a basic template.
+    """
+    request = rf.get("/")
+    request.LANGUAGE_CODE = "en"
 
-    # Set the language without categories
-    with translation.override("de"):
-        rendered = article_categories()
-        assert len(rendered["categories"]) == 0
+    picsou = CategoryFactory(title="Picsou", slug="picsou", language="en")
+    donald = CategoryFactory(title="Donald", slug="donald", language="en")
+    CategoryFactory(title="Dupont", slug="dupont", language="fr")
 
+    template_to_render = Template("{% load lotus %}{% get_categories_html picsou %}")
+    rendered_template = template_to_render.render(
+        Context({"request": request, "picsou": picsou})
+    )
 
-def test_categories_ordering(db):
-    en_categories = CategoryFactory.create_batch(5, language="en")
-
-    with translation.override("en"):
-        rendered = article_categories()
-        assert len(rendered["categories"]) == 5
-
-        rendered_titles = [category["title"] for category in rendered["categories"]]
-        original_titles = sorted([category.title for category in en_categories])
-
-        # Verify that the categories are ordered by title
-        assert rendered_titles == original_titles
-
-
-def test_template_rendering(db):
-    en_categories = CategoryFactory.create_batch(3, language="en")
-
-    with translation.override("en"):
-        categories = article_categories()
-        context = Context(categories)
-        template_to_render = Template(
-            "{% load lotus %}{% news_categories %}"
-        )
-        rendered_template = template_to_render.render(context)
-
-        for category in en_categories:
-            assert f'href="{category.get_absolute_url()}"' in rendered_template
-            assert category.title in rendered_template
+    # Expected items are in built HTML
+    dom = html_pyquery(rendered_template)
+    assert sorted([
+        (a.text.strip(), a.get("href"), a.get("aria-current"))
+        for a in dom.find("a")
+    ]) == [
+        ("Donald", donald.get_absolute_url(), None),
+        ("Picsou", picsou.get_absolute_url(), "page"),
+    ]